Update package key manager

+++ src/main/java/onqlave_key_manager/types/KeyData.java
@@ -9,5 +9,5 @@ public interface KeyData {
 
     KeyMaterialType getKeyMaterialType();
 
-    long getVersion();
+    int getVersion();
 }
